requirements.txt
Flask==2.2.5
PyYAML==6.0
requests==2.31.0
langdetect==1.0.9
gunicorn==20.1.0
